+++ web/wiki/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.22.3] - 2024-07-30
+
+### Fixed
+
+- Fixed Login and tests to pass tests at Login.test.tsx
+- Changed openapiFetcher file to manage openapi error
+
 ## [0.22.2] - 2024-07-25
 
 ### Fixed